Fix #2
Guh, why isn't SQLite even a little bit type-safe?

The underlying problem, as suggested in @storyboyle's bug report, is
that we were inserting todos into the database with a `done` value of
the SQL *string* `'false'`, rather than the boolean value. This was
mucking with the sorting order of items until the first time their
done-value was changed.

This patch fixes the insert code. It does *not* attempt to repair
databases created by older versions of the app; while it would be easy
enough to just add an SQL call like:

    UPDATE todos SET done = 0 WHERE done = 'false'

To run on startup, it's an extra bit of logic that would have to stay
there indefinitely. The issue is easily fixed by the user on an existing
grain, and to my knowledge there are basically two users right now
(myself and @storyboyle), so it seems silly to leave that code there
potentially forever. I've decided it isn't worth adding code to fix
existing databases.