Add a small program for re-sizing images.
I'm going to use this to make larger versions of the icon. Right now
we're using the same 14x14 image for all icons, including the large
ones. However, my intent was for the larger images to still appear
low-resolution, and image viewers seem to try to interpolate the pixels.
So, we'll resize the images ourselves so we get sharp square pixels.